+++ api/citiesModels/cities.js
@@ -17,6 +17,14 @@ class User {
         return user;
     }
 
+    static async userQueryByUsername(username) {
+        const query = new Parse.Query(Parse.User);
+        query.equalTo("username", username);
+        const user = await query.find();
+        await user.fetch();
+        return user;
+    }
+
     static async dereferencePointers(pointers) {
         let following = [];
         // Used for loop to iterate over the User pointers because the map() function is
@@ -25,7 +33,7 @@ class User {
         for (const pointer of pointers) {
             const query = new Parse.Query(Parse.User);
             const user = await query.get(pointer["id"]);
-            following.push(user.get("username"));
+            following.push(user);
         }
         return following;
     }
