+++ api/index.js
@@ -3,11 +3,9 @@ const morgan = require('morgan')
 const cors = require('cors')
 const Parse = require('parse/node')
 const authRoute = require('./authentication/authentication')
-const citiesRoute = require('./citiesRoutes/cities')
+
 const usersRoute = require('./userRoutes/users')
-//import Parse from 'parse/dist/parse.min.js';
 const {PARSE_APP_ID, PARSE_JAVASCRIPT_KEY} = require('./config')
-//const { default: App } = require('../ui/src/components/App/App')
 
 const app = express()
 const port = process.env.PORT || 3001
@@ -16,7 +14,7 @@ app.use(express.json())
 app.use(morgan("tiny"))
 app.use(cors())
 app.use('/auth', authRoute)
-app.use('/cities', citiesRoute)
+
 app.use('/users', usersRoute)
 
 // initalize Parse
