+++ api/userRoutes/users.js
@@ -1,7 +1,21 @@
 const express = require("express");
 const router  = express.Router();
-const User  = require("../models/models")
+const User  = require("../citiesModels/cities")
 const Parse = require('parse/node')
+
+
+// Get list of favorited cities
+router.get("/get-cities", async (req, res, next) => {
+    try {
+        const user = await User.getUser();
+        let cities = user.get("cities");
+        res.status(200).json({ cities })
+    } catch (err) {
+        next(err);
+    }
+})
+
+
 
 // Get list of favorited users
 router.get("/get-users", async (req, res, next) => {
@@ -14,11 +28,17 @@ router.get("/get-users", async (req, res, next) => {
     }
 })
 
+router.get("/get-friends-cities", async (req, res, next) => {
+    const friend = await User.userQueryByUsername(req.body.user);
+    console.log(friend.cities)
+    const cities = friend.cities;
+    res.status(200).json({ cities });
+})
+
 router.get("/get-following", async (req, res, next) => {
     try {
         let currentUser = await User.getUser();
         let followingPointers = currentUser.get("friends");
-
         let following = await User.dereferencePointers(followingPointers)
         res.status(200).json({  following })
     } catch (err) {
@@ -77,5 +97,37 @@ router.post('/remove-friend', async (req, res, next) => {
     }
 });
 
+// Add route adds a city to the end of the "cities" array in the User object
+router.post('/add-city', async (req, res, next) => {
+    try {
+        const user = await User.getUser();
+        if (user == null) {
+            res.sendStatus(400)
+        }
+        user.addUnique("cities", req.body.city)
+        console.log(req.body.city)
+        await user.save()
+        res.sendStatus(200)
+    } catch(err) {
+        next(err)
+    }
+});
+
+// Remove route removes a city from the end of the "cities" array in
+// the User object
+router.post('/remove-city', async (req, res, next) => {
+    try {
+        const user = await User.getUser();
+
+        if (user == null) {
+            res.sendStatus(400)
+        }
+        user.remove("cities", req.body.city)
+        await user.save()
+        res.sendStatus(200)
+    } catch(err) {
+        next(err)
+    }
+});
 
 module.exports = router;
